refactor: refactor the firewall.go (#32)
